# src/swagger_files/QuickCode.Demo.SmsManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Demo - SmsManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/sms-manager-module/info-messages/count": {
      "get": {
        "tags": [
          "InfoMessages"
        ],
        "operationId": "api-sms-manager-module-info-messages-count-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sms-manager-module/info-messages/{id}": {
      "get": {
        "tags": [
          "InfoMessages"
        ],
        "operationId": "api-sms-manager-module-info-messages--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoMessagesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "InfoMessages"
        ],
        "operationId": "api-sms-manager-module-info-messages--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfoMessagesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InfoMessagesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InfoMessagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "InfoMessages"
        ],
        "operationId": "api-sms-manager-module-info-messages--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sms-manager-module/info-types/{infoTypesId}/info-messages/{infoMessagesId}": {
      "get": {
        "tags": [
          "InfoTypes"
        ],
        "operationId": "api-sms-manager-module-info-types--infoTypesId--info-messages--infoMessagesId--get",
        "parameters": [
          {
            "name": "infoTypesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "infoMessagesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoTypesInfoMessagesKeyRestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sms-manager-module/sms-senders/{smsSendersId}/info-messages/{infoMessagesId}": {
      "get": {
        "tags": [
          "SmsSenders"
        ],
        "operationId": "api-sms-manager-module-sms-senders--smsSendersId--info-messages--infoMessagesId--get",
        "parameters": [
          {
            "name": "smsSendersId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "infoMessagesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsSendersInfoMessagesKeyRestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InfoMessagesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "infoTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "InfoTypesInfoMessagesKeyRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "infoTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SmsSendersInfoMessagesKeyRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "infoTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}